+++ packages/select/src/Component.test.tsx
@@ -1,4 +1,5 @@
-import React, { MutableRefObject } from 'react';
+import React, { ForwardRefRenderFunction, MutableRefObject } from 'react';
+
 import { fireEvent, render, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
@@ -9,6 +10,10 @@ import { FieldProps as BaseFieldProps, Select, OptionsListProps, OptionProps } f
 import * as fieldModule from './components/field';
 import * as optionsListModule from './components/options-list';
 import * as optionModule from './components/option';
+
+type PopoverComponent = {
+    render?: ForwardRefRenderFunction<HTMLDivElement, popoverModule.PopoverProps>;
+};
 
 const baseProps = {
     options: [],
@@ -509,7 +514,8 @@ describe('Select', () => {
         });
 
         it('should transfer props to Popover', () => {
-            const spy = jest.spyOn(popoverModule, 'Popover');
+            const PopoverComponent = popoverModule.Popover as PopoverComponent;
+            const spy = jest.spyOn(PopoverComponent, 'render');
 
             const cb = () => undefined;